Added another common block, with mismatched usage.
The common block is not identical in the two source files, a.f90 and b.f90.
The compiler emits the following assembly code:

==> a.s <==
  ret
  .cfi_endproc
  .LFE0:
  .size   another_print_, .-another_print_
  .comm   more_2_,4,16
  .comm   more_,4,16
  .comm   mismatch_,8,16
  .comm   factors_,24,16

==> b.s <==
  ret
  .cfi_endproc
  .LFE0:
  .size set_data_, .-set_data_
  .comm more_2_,4,16
  .comm more_,4,16
  .comm mismatch_,4,16

The common block has different sizes, 4 and 8 bytes, respectively, and the
linker will merge the two, using the larger size. The third argument specifies
byte alignment in ELF format object files.